HUD rework: stats panel, incremental saves, death linger, XP as long, nav layout
- Simplify stats panel: 4 stats (Vitality, Strength, Agility, Spirit) in 2x2
  grid with inline PlusButton icons (unpressed/pressed/disabled states)
- Remove Mana allocation, XP display from equipment panel; show Class + Level
- Combat stats: single column with full names (Attack, Defence, etc.)
- Incremental save on every enemy kill and item pickup (delta tracking)
- Enemy death linger: 3s corpse visibility after death animation
- Change XP from int to long for large value support
- Display full XP value (e.g. "1234 / 5000") above EXP orb
- Reduce respawn timer from 120s to 20s
- Move nav buttons above skill slots, remove kill counter
- Fix orb fill circles to match orb sprite size (remove mask inset)
- Add UIAtlasHelper, UISpriteAnimator, OrbMaskCircle sprite
- Update atlas with PlusButton sprites and new UI art
- Various scene and editor setup improvements

+++ Assets/_Game/Core/Character/CharacterState.cs
@@ -4,7 +4,7 @@ namespace ConquerChronicles.Core.Character
     {
         public CharacterClass Class;
         public int Level;
-        public int XP;
+        public long XP;
         public int CurrentHP;
         public int CurrentMP;
         public int StatPointsAvailable;
